# core-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Национальный сервис мониторинга цен на продовольственные товары(OpenAPI 3.0)
  description: |-
    Национальный сервис мониторинга цен на продовольственные товары
    
    Полезные ссылки:
    - [Репозиторий](https://github.com/AnnaDesert/Learning)
  version: 1.0.11
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /user:
    get:
      tags:
        - user
      summary: Получить всеx пользователей
      description: Получить всеx пользователей
      operationId: allUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags:
        - user
      summary: Добваить пользователя
      description: Добваить нового пользователя
      operationId: singUpUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /user/{id}:
    get:
      tags:
        - user
      summary: Получить пользователя по ID
      description: Получить пользователя по ID
      operationId: userById
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
    delete:
      tags:
        - user
      summary: Удалить пользователя по ID
      description: Удалить пользователя по ID
      operationId: removeUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - user
      summary: Изменить пользователя
      description: Изменить пользователя
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /user/phoneNumber/{phoneNumber}:
    get:
      tags:
        - user
      summary: Получить пользователя по номеру телефона
      description: Получить пользователя по номеру телефона
      operationId: userByPhone
      parameters:
        - name: phoneNumber
          in: path
          description: Номер телефона
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
  /user/name/{name}:
    get:
      tags:
        - user
      summary: Получить пользователя по имени
      description: Получить пользователя по имени
      operationId: userByName
      parameters:
        - name: name
          in: path
          description: Имя пользователя
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
  /user/email/{email}:
    get:
      tags:
        - user
      summary: Получить пользователя по адресу электронной почты
      description: Получить пользователя по адресу электронной почты
      operationId: userByEmail
      parameters:
        - name: email
          in: path
          description: Адрес электронной почты
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
  /shop:
    get:
      tags:
        - shop
      summary: Получить все торговые точки
      description: Получить все торговые точки
      operationId: allShop
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
    post:
      tags:
        - shop
      summary: Добваить торговую точку
      description: Добваить торговую точку
      operationId: addShop
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /shop/{id}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по ID
      description: Получить торговую точку по ID
      operationId: getShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ShopDTO'
    delete:
      tags:
        - shop
      summary: Удалить торговую точку по ID
      description: Удалить торговую точку по ID
      operationId: removeShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - shop
      summary: Изменить торговую точку
      description: Изменить торговую точку
      operationId: updateShop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /shop/name/{name}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по названию
      description: Получить торговую точку по названию
      operationId: getShopByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
  /shop/address/{address}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по адресу
      description: Получить торговую точку по адресу
      operationId: getShopByAddress
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
...
components:
  schemas:
    UserCreationDTO:
      required:
        - email
        - name
        - password
        - phoneNumber
        - role
        - surname
      type: object
      properties:
        name:
          type: string
        password:
          type: string
        surname:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
    ShopCreationDTO:
      required:
        - address
        - idAdmin
        - name
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
...
    UserDTO:
      required:
        - email
        - id
        - name
        - phoneNumber
        - role
        - surname
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
    ShopDTO:
      required:
        - address
        - id
        - idAdmin
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
...